+++ Runtime/TomlSerializer.Deserialize.cs
@@ -96,9 +96,7 @@ namespace UnderLogic.Serialization.Toml
                 throw new InvalidOperationException($"Type {type.Name} is not serializable");
 
             // Allow the top-level object to specify a default field casing
-            var objectFieldCasing = StringCasing.Default;
-            if (TryGetAttribute<TomlCasingAttribute>(type, out var objectCasingAttribute))
-                objectFieldCasing = objectCasingAttribute.Casing;
+            var objectFieldCasing = GetStringCasing(type);
 
             var fields = type.GetFields(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
 
@@ -117,8 +115,7 @@ namespace UnderLogic.Serialization.Toml
                 else
                 {
                     // Allow the field to specify a custom field casing
-                    if (TryGetAttribute<TomlCasingAttribute>(field, out var fieldCasingAttribute))
-                        fieldCasing = fieldCasingAttribute.Casing;
+                    fieldCasing = GetStringCasing(field, objectFieldCasing);
 
                     // Apply the field casing to the key
                     if (fieldCasing != StringCasing.Default)