+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.6] - 2023-01-04
+
+### Added
+
+- `TomlCamelCaseAttribute` custom attribute
+- `TomlPascalCaseAttribute` custom attribute
+- `TomlSnakeCaseAttribute` custom attribute (allow for uppercase)
+- `TomlKebabCaseAttribute` custom attribute
+
+### Removed
+
+- `TomlCasingAttribute` custom attribute in favor of specific ones
+
 ## [0.4.5] - 2023-01-04
 
 ### Added